Refactor navbar to the renderable parts pattern
