--- src/components/navbar/NavbarLabel.css
@@ -1,7 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-.NavbarLabel {
-  font-size: 14px;
-  line-height: 1;
-}
-
