--- src/components/navbar/NavbarLabel.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const NavbarLabel = () =>
-  <h2 className="NavbarLabel">Ello</h2>
-
-export default NavbarLabel
-
