--- src/components/navbar/NavbarLink.css
@@ -1,121 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-.NavbarLinks {
-  transition: transform var(--speed) var(--ease);
-}
-
-.NavbarLink {
-  position: relative;
-  height: 40px;
-  overflow: hidden;
-  font-size: 14px;
-  line-height: 40px;
-  color: #aaa;
-  white-space: nowrap;
-  vertical-align: middle;
-  transition: width var(--speed) var(--ease), color var(--speed) ease, background-color var(--speed) ease;
-}
-
-.NavbarLink::before {
-  @apply --hitarea;
-}
-
-.NavbarLink.isSignUp {
-  width: 80px;
-  color: #fff;
-  text-align: center;
-  background-color: #00d100;
-  border-radius: 5px;
-}
-
-.NavbarLink > .SVGIcon {
-  margin-top: -5px;
-}
-
-.NavbarLinkLabel {
-  position: relative;
-  display: inline-block;
-  vertical-align: middle;
-}
-
-.NavbarLinkLabel::after {
-  position: absolute;
-  bottom: 8px;
-  left: 0;
-  width: 0;
-  height: 2px;
-  content: "";
-  background-color: transparent;
-  transition: width var(--speed) var(--ease), background-color var(--speed) ease;
-}
-
-.NavbarLink.isSignUp .NavbarLinkLabel::after {
-  display: none;
-}
-
-/* Active and hover states
- * @see NavbarIcons for icon animations and coloring
-------------------------------------- */
-
-.NavbarLink.isActive > .NavbarLinkLabel,
-.NavbarLink:active > .NavbarLinkLabel,
-.no-touch .NavbarLink:hover > .NavbarLinkLabel {
-  color: #000;
-}
-
-.NavbarLink.isSignUp.isActive,
-.NavbarLink.isSignUp:active,
-.no-touch .NavbarLink.isSignUp:hover {
-  background-color: #00b100;
-}
-
-/* stylelint-disable selector-max-specificity */
-.NavbarLink.isSignUp.isActive > .NavbarLinkLabel,
-.NavbarLink.isSignUp:active > .NavbarLinkLabel,
-.no-touch .NavbarLink.isSignUp:hover > .NavbarLinkLabel {
-  color: #fff;
-}
-
-.NavbarLink.isActive > .NavbarLinkLabel::after,
-.NavbarLink:active > .NavbarLinkLabel::after,
-.no-touch .NavbarLink:hover > .NavbarLinkLabel::after {
-  width: 100%;
-  background-color: currentColor;
-}
-
-.NavbarLink.isNotificationsUnread > .NavbarLinkLabel::before {
-  position: absolute;
-  top: 5px;
-  left: -5px;
-  width: 5px;
-  height: 5px;
-  content: "";
-  background-color: var(--color-danger);
-  border-radius: 50%;
-}
-
-.no-touch .NavbarLink.isNotificationsUnread:hover > .NavbarLinkLabel::before {
-  left: -7px;
-  width: 7px;
-  height: 7px;
-}
-/* stylelint-enable selector-max-specificity */
-
-/* Stream button has an acceptable Avatar being dragged - turn background yellow */
-[data-dragging-priority="noise"] .NavbarLink[href="/following"],
-[data-dragging-priority="friend"] .NavbarLink[href="/starred"],
-[data-dragging-priority="inactive"] .NavbarLink[href="/following"],
-[data-dragging-priority="inactive"] .NavbarLink[href="/starred"] {
-  padding-right: 10px;
-  padding-left: 10px;
-  background-color: #ffc;
-}
-
-/* Stream button has an acceptable Avatar being dragged over it - turn background green */
-[data-dragging-priority="noise"] .NavbarLink.hasDragOver[href="/following"],
-[data-dragging-priority="friend"] .NavbarLink.hasDragOver[href="/starred"],
-[data-dragging-priority="inactive"] .NavbarLink.hasDragOver[href="/following"],
-[data-dragging-priority="inactive"] .NavbarLink.hasDragOver[href="/starred"] {
-  background-color: #cfc;
-}
-
