--- src/components/navbar/NavbarMark.css
@@ -1,15 +1,0 @@
-@import "../../lib/styles/properties.css";
-
-.NavbarMark {
-  position: relative;
-  transform: scale(0.75);
-  transform-origin: top left;
-}
-
-@media (--break-2) {
-  .NavbarMark {
-    transform: none;
-    transform-origin: center center;
-  }
-}
-
