+++ src/main.css
@@ -61,14 +61,9 @@
 
 @import "components/modals/Modal.css";
 
-@import "components/navbar/Navbar.css";
 @import "components/navbar/NavbarIcons.css";
-@import "components/navbar/NavbarLabel.css";
-@import "components/navbar/NavbarLink.css";
-@import "components/navbar/NavbarMark.css";
-@import "components/navbar/NavbarMorePostsButton.css";
-@import "components/navbar/NavbarOmniButton.css";
-@import "components/navbar/NavbarProfile.css";
+@import "components/navbar/NavbarParts.css";
+@import "components/navbar/NavbarRenderables.css";
 
 @import "components/navigation/TreeButton.css";
 @import "components/navigation/TreePanel.css";
